% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{cleanDoi}
\alias{cleanDoi}
\title{cleanDoi}
\usage{
cleanDoi(doi)
}
\arguments{
\item{doi}{a doi string}
}
\value{
the values of the node
}
\description{
cleanDoi
}
\examples{
cleanDoi("https://doi.org/10.1212/01.wnl.0000260060.60870.89")

}
