% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{getContentByGetLink}
\alias{getContentByGetLink}
\title{getContentByGetLink}
\usage{
getContentByGetLink(link, waitTime = 0.3)
}
\arguments{
\item{link}{a string of characters}

\item{waitTime}{a number. Waiting of the program}
}
\value{
a string of characters of the returned content
}
\description{
getContentByGetLink
}
\examples{
link <- "https://www.ebi.ac.uk/europepmc/webservices/rest/search?query=pinkeye"
content <- getContentByGetLink(link, 0.3)

}
