% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{getContentByPostLink}
\alias{getContentByPostLink}
\title{getContentByPostLink}
\usage{
getContentByPostLink(link, waitTime = 0.3)
}
\arguments{
\item{link}{a string of characters}

\item{waitTime}{a number. Waiting of the program}
}
\value{
a string of characters of the returned content
}
\description{
getContentByPostLink
}
\examples{
link <- getAPIlink(query = "pinkeye")
content <- getContentByPostLink(link, 0.3)

}
