% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{getJson}
\alias{getJson}
\title{getJson}
\usage{
getJson(
  query,
  endpoint = "search",
  resultType = "lite",
  synonym = "FALSE",
  cursorMark = "*",
  pageSize = "1",
  callback = "",
  waitTime = 0.1
)
}
\arguments{
\item{query}{a string of characters. Entrez text query.}

\item{endpoint}{a string of characters, valid endpoint of NCBI Utilities, e.g. "search",}

\item{resultType}{a string of characters.}

\item{synonym}{a string of characters.}

\item{cursorMark}{a string of characters.}

\item{pageSize}{a string of characters.}

\item{callback}{a string of characters.}

\item{waitTime}{a number to wait in each call}
}
\value{
a XMLInternalDocument
}
\description{
getJson
}
\examples{
getJson(query = "pinkeye")


}
