% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{retriveXmlNodeValuefromDoc}
\alias{retriveXmlNodeValuefromDoc}
\title{retriveXmlNodeValuefromDoc}
\usage{
retriveXmlNodeValuefromDoc(doc, nodePosition)
}
\arguments{
\item{doc}{the parsed XML file}

\item{nodePosition}{the node position of the xml file}
}
\value{
the values of the node
}
\description{
retriveXmlNodeValuefromDoc
}
\examples{
 doc <- getDoc("pinkeye")
retriveXmlNodeValuefromDoc(doc, "//pmid")

}
