% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europeanPMC_methods.R
\name{search_pmc}
\alias{search_pmc}
\title{search_pmc}
\usage{
search_pmc(
  query,
  resultType = "lite",
  synonym = "FALSE",
  callback = "",
  waitTime = 0.1
)
}
\arguments{
\item{query}{a string of character}

\item{resultType}{a string of character}

\item{synonym}{a string of character}

\item{callback}{a string of character}

\item{waitTime}{a number}
}
\description{
search_pmc
}
\examples{
search_pmc("pinkeye")

}
